% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_batch.R
\name{cxlib_batch}
\alias{cxlib_batch}
\title{Simple utility to execute an R program script}
\usage{
cxlib_batch(
  x,
  options = list(logs = NULL, log.fileext = "Rout"),
  silent = FALSE
)
}
\arguments{
\item{x}{A job}

\item{Rout}{Execution log file}

\item{user.libary}{Enable user library}

\item{overwrite}{Enable overwriting existing output}

\item{overwrite.input}{Enable overwriting inputs}

\item{libs}{(Experimental) A vector of library paths}

\item{slient}{(Not implemented) Disable messaging}
}
\value{
Named list of execution references
}
\description{
A job is a list of tasks to be executed. In its simplest form, it one or more
programs that executes in batch mode.

A job is executed in an isolated, temporary and transient working area.

Note: All paths are relative to the current R session working directory, i.e.
\code{getwd()}.
#'
Program required inputs and output locations are annotated in the program file
using standard annotation syntax anywhere in the program file. Each annotation
is defined on a separate line starting with the comment character \verb{#}. One or
more spaces separate the annotation keyword and its value.

Program input is specified using \verb{@input} followed by the input directory or
file path. If the input is a directory, all files in that directory is
included.

Program output locations is specified using the \verb{@output} directive. It is
assumed that the output location is a directory.

If the \code{log.file} is specified as a directory, a log file created will use
the convention \if{html}{\out{<program name>}}.Rout. To use a different file name convention,
specify the \code{log.file} as a full file name. The parent directory to the log
file must exist.

The working directory is either set to the parent directory of the program
\code{use.wd = "program"} or the root of the batch execution area \verb{use.wd = "root}.
The default is the root of the batch execution area

The user R package library can be disabled with \code{user.library = FALSE}. If
\code{internal = TRUE}, user library is disabled by default.

The \code{internal} mode is used when \code{cxlib_batch()} is used as an internal
utility.

If \code{silent=FALSE} the program execution log is printed to the console.
}
