% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_propertiesread.R
\name{cxlib_propertiesread}
\alias{cxlib_propertiesread}
\title{Utility to read properties}
\usage{
cxlib_propertiesread(x)
}
\arguments{
\item{x}{A vector of properties file paths}
}
\value{
A character vector of named property elements
}
\description{
A property file contains key/value pair entries delimited by an
equal sign or colon (\code{:}).

The first existing properties file in the vector of paths \code{x} is read.

The property key or name is the string before the first equal sign or colon.
The value is the remainder of the line after the first equal sign or colon.
Any equal signs after the first occurrence is considered part of the value.

A property name includes the characters A-Z case insensitive, digits 0-9 and
punctuation period and underscore.

The last occurrence of a property definition is used. Prior definitions
are ignored.

A line that starts with a hash (\verb{#}) or exclamation (\code{!}) is assumed a
comment and is ignored.

Trailing and leading white space is ignored.
}
\examples{

# not run
result <- cxlib_propertiesread( "/my/path/to/my.properties" )

}
