% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_batchlocalfs_publish.R
\name{.cxlib_batch_localfs_publish}
\alias{.cxlib_batch_localfs_publish}
\title{Internal utility to publish results of a job to local file system}
\usage{
.cxlib_batch_localfs_publish(x)
}
\arguments{
\item{x}{Job results}
}
\value{
Invisible Job results
}
\description{
Internal utility to publish results of a job to local file system
}
\keyword{internal}
