% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_batchlocalfs_stage.R
\name{.cxlib_batch_localfs_stage}
\alias{.cxlib_batch_localfs_stage}
\title{Internal utility function to stage a local program with inputs}
\usage{
.cxlib_batch_localfs_stage(x, options = NULL)
}
\arguments{
\item{x}{A character vector with the program paths}

\item{options}{A list of options}
}
\value{
A list of named entries detailing the staged content
}
\description{
The vector of programs \code{x} is assumed to define the order of execution
}
\keyword{internal}
