% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_programannotations.R
\name{.cxlib_programannotations}
\alias{.cxlib_programannotations}
\title{Retrieve annotations from a given program}
\usage{
.cxlib_programannotations(x)
}
\arguments{
\item{x}{A program path}
}
\value{
A named list of annotations
}
\description{
An annotation is a special comment that follows the annotation convention.

One annotation per line. Wrapping an annotation across several lines is not
supported.

Annotation follows the convention of first non-space character on a line is
a \verb{#} followed by an \code{@} and the annotation keyword, e.g. \verb{# @cx.input ...}.

Note: Only annotations with the keyword prefix \verb{cx.<word>} are processed and
the return drops the \code{cx.} prefix in annotation names.

There is at least one white space character between the annotation keyword
and value. The value is terminated by end of line character. Leading and
trailing white space is removed.
}
\keyword{internal}
