% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_programexec.R
\name{.cxlib_programexec}
\alias{.cxlib_programexec}
\title{Internal utility function to execute a R program in batch}
\usage{
.cxlib_programexec(x, work.area = NULL)
}
\arguments{
\item{x}{Program file}

\item{work.area}{Working area for program execution}
}
\value{
A list of the execution results
}
\description{
Internal utility function to execute a R program in batch
}
\keyword{internal}
