% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_referenceid.R
\name{.cxlib_referenceid}
\alias{.cxlib_referenceid}
\title{Internal utility function to generate a unique reference iD of different types}
\usage{
.cxlib_referenceid(type = NULL)
}
\arguments{
\item{type}{ID type}
}
\value{
A character vector with the reference ID
}
\description{
Type equal to \code{uuid} returns a secure random UUID

Type equal to \code{sha} or \code{raw} returns the SHA-1 of the UUID

Type equal to \code{NULL} returns formatted SHA-1 of the UUID
}
\keyword{internal}
