% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_standardpath.R
\name{.cxlib_standardpath}
\alias{.cxlib_standardpath}
\title{Simple utility to standard path reference}
\usage{
.cxlib_standardpath(x, normalize = FALSE)
}
\arguments{
\item{x}{A vector of paths}

\item{normalize}{Normalize path}
}
\value{
A vector of paths in standard reference format
}
\description{
A standard reference uses single forward slashes as directory and file
separator.
}
\keyword{internal}
