% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxlib_batch.R
\name{cxlib_batch}
\alias{cxlib_batch}
\title{Simple utility to execute an R program script}
\usage{
cxlib_batch(
  x,
  options = list(logs = NULL, log.fileext = "Rout"),
  silent = FALSE
)
}
\arguments{
\item{x}{A job}

\item{options}{List of options for executing R programs}

\item{silent}{Disable messaging}
}
\value{
Named list of execution results
}
\description{
A job is a list of tasks to be executed. In its simplest form, it one or more
programs that executes in batch mode.

A job is executed in an isolated, temporary and transient working area.

Note: All paths are relative to the current R session working directory, i.e.
\code{getwd()}.

Program required inputs and output locations are annotated in the program file
using standard annotation syntax anywhere in the program file. Each annotation
is defined on a separate line starting with the comment character \verb{#}. One or
more spaces separate the annotation keyword and its value.

Program input is specified using \verb{@cx.input} followed by the input directory or
file path. If the input is a directory, all files in that directory is
included.

Program output locations is specified using the \verb{@cx.output} annotation. It is
assumed that the output location is a directory.

Only files created, updated or deleted in the output directory are returned
in addition to the program log.

The function supports different execution options.

The \code{logs} option is the directory where logs are stored. If the option is
\code{NULL}, the program directory is used.

The \code{log.fileext} specifies te log file extension or suffix. Default is the
standard \code{Rout}.
}
