% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxlib_programexec.R
\name{.cxlib_programexec}
\alias{.cxlib_programexec}
\title{Internal utility function to execute a R program in batch}
\usage{
.cxlib_programexec(x, job.id = NULL, options = NULL)
}
\arguments{
\item{x}{Program file}

\item{job.id}{Jon ID reference}

\item{options}{Execution options}
}
\value{
A list of the execution results
}
\description{
\code{options} is a named list of options when executing the program.

Option \code{work.area} specifies the work area to use when executing the program.
If not specified, the current working directory is used.

Option \code{log} specifies an alternative log file. If the \code{log} option is not
specified, the log file name is derived from the program file name with
suffix \code{.Rout}and is saved to the program parent directory.
}
\keyword{internal}
